refactor: split handlers.go into domain-specific files (#279)
* Trigger deploy: setup.sh run, secrets provisioned

* refactor: split handlers.go into domain-specific files

handlers.go (3787 lines, 87 methods) split into 9 files by domain:
- handlers_branches.go  — branches, rebase, auto-merge, branch status, agent context
- handlers_proposals.go — proposals, reviews, checks, merge, review comments
- handlers_issues.go    — issues, comments, refs, mention parsing
- handlers_orgs.go      — orgs, membership, invites
- handlers_files.go     — files, tree, diff, commits, archive, chain
- handlers_events.go    — subscriptions, SSE streaming
- handlers_repos.go     — repo CRUD, purge
- handlers_releases.go  — releases
- handlers_users.go     — role management
- handlers.go           — dispatcher, parseRepoPath, writeError, validateRepo

No handler logic changed; pure reorganization.


---------
